+++ resources/views/emails/hr/bwr-id-document-auto-deleted.blade.php
@@ -1,6 +1,8 @@
 {{-- SPDX-FileCopyrightText: 2026 SecPal Contributors --}}
 {{-- SPDX-License-Identifier: AGPL-3.0-or-later --}}
 
+@php($employeeDetailUrl = rtrim((string) (config('app.frontend_url') ?: url('/')), '/') . '/employees/' . $employee->id)
+
 <x-mail::message>
 # {{ __('ID Document Copy Deleted After BWR Approval') }}
 
@@ -20,7 +22,7 @@
 
 **{{ __('Legal Basis') }}:** {{ $legalBasis }}
 
-<x-mail::button :url="url('/admin/employees/' . $employee->id)">
+<x-mail::button :url="$employeeDetailUrl">
 {{ __('View Employee Details') }}
 </x-mail::button>